Added taskbar/tasklist.
All running apps are displayed when Ignis is started.
When new window/app is opened then an app icon is added to the tasklist.
When window/app is closed then an app icon is removed from tasklist.
When app icon is clicked then the focus is moved to given app window.